chore: use NewReplacer for newline substitution
