+++ importer_wizard_template.go
@@ -4,6 +4,9 @@ import (
 	"strings"
 
 	tea "github.com/charmbracelet/bubbletea"
+	"github.com/charmbracelet/x/ansi"
+
+	"github.com/marang/goemqutiti/ui"
 )
 
 // updateTemplate handles the template entry step.
@@ -21,4 +24,16 @@ func (w *ImportWizard) updateTemplate(msg tea.Msg) (tea.Model, tea.Cmd) {
 		}
 	}
 	return w, cmd
+}
+
+// viewTemplate renders the topic template step.
+func (w *ImportWizard) viewTemplate(bw, wrap int) string {
+	names := make([]string, len(w.headers))
+	for i, h := range w.headers {
+		names[i] = "{" + h + "}"
+	}
+	help := "Available fields: " + strings.Join(names, " ")
+	help = ansi.Wrap(help, wrap, " ")
+	content := w.tmpl.View() + "\n" + help + "\n[enter] continue  [ctrl+n] next  [ctrl+p] back"
+	return ui.LegendBox(content, "Topic Template", bw, 0, ui.ColBlue, true, -1)
 }
